feat: add `ak.testing.arrays_approx_equal`
